% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amt_with_frequency.R
\name{amt_every_n_weeks}
\alias{amt_every_n_weeks}
\title{Amount every n weeks}
\usage{
amt_every_n_weeks(start_date, amount, n = 1L, end_date = as.Date(NA))
}
\arguments{
\item{start_date}{Start date}

\item{amount}{Amount}

\item{n}{Number of weeks}

\item{end_date}{End date}
}
\description{
Amount every n weeks
}
