% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_balance.R
\name{simulate_balance}
\alias{simulate_balance}
\title{Simulate balance}
\usage{
simulate_balance(initial_amount, end_date, ...)
}
\arguments{
\item{initial_amount}{Initial amount}

\item{end_date}{End date}

\item{...}{Named calls to \verb{amt_*} functions}
}
\description{
Simulate balance
}
